EUI-4549 Global Search 16 digit case reference number in menu header unhappy path
