+++ src/app/components/hmcts-global-header/hmcts-global-header.component.ts
@@ -1,23 +1,22 @@
-import { AfterViewInit, Component, ElementRef, EventEmitter, Input, OnChanges, OnDestroy, Output, SimpleChanges, ViewChild, ViewChildren, Renderer2 } from '@angular/core';
+import { Component, EventEmitter, Input, OnChanges, Output, SimpleChanges } from '@angular/core';
 import { FeatureToggleService } from '@hmcts/rpx-xui-common-lib';
 import { Store } from '@ngrx/store';
-import { BehaviorSubject, Observable, of, Subscription } from 'rxjs';
+import { BehaviorSubject, Observable, of } from 'rxjs';
 import { map, switchMap, tap } from 'rxjs/operators';
 import * as fromNocStore from '../../../noc/store';
+import { SearchStatePersistenceKey } from '../../../search/enums';
+import { SearchParameters } from '../../../search/models';
+import { SearchService } from '../../../search/services/search.service';
 import { NavItemsModel } from '../../models/nav-item.model';
 import { UserNavModel } from '../../models/user-nav.model';
 import { UserService } from '../../services/user/user.service';
-import { SearchService } from '../../../search/services/search.service';
-import { ActivatedRoute, Router } from '@angular/router';
-import { NoResultsMessageId, SearchStatePersistenceKey } from '../../../search/enums';
-import { SearchParameters } from '../../../search/models';
 
 @Component({
     selector: 'exui-hmcts-global-header',
     templateUrl: './hmcts-global-header.component.html',
     styleUrls: ['./hmcts-global-header.component.scss']
 })
-export class HmctsGlobalHeaderComponent implements OnChanges, AfterViewInit, OnDestroy {
+export class HmctsGlobalHeaderComponent implements OnChanges {
 
   @Input() public set showNavItems(value: boolean) {
     this.showItems = value;
@@ -33,8 +32,6 @@ export class HmctsGlobalHeaderComponent implements OnChanges, AfterViewInit, OnD
   @Input() public decorate16DigitCaseReferenceSearchBox: boolean;
   @Output() public navigate = new EventEmitter<string>();
 
-  @ViewChildren('caseReference') public caseReferenceEl: ElementRef;
-
   public showItems: boolean;
   public userValue = true;
   public tab;
@@ -44,7 +41,6 @@ export class HmctsGlobalHeaderComponent implements OnChanges, AfterViewInit, OnD
   public get rightItems(): Observable<NavItemsModel[]> {
     return this.menuItems.right.asObservable();
   };
-  public searchSubscription$: Subscription;
 
   private menuItems = {
     left: new BehaviorSubject<NavItemsModel[]>([]),
@@ -55,10 +51,7 @@ export class HmctsGlobalHeaderComponent implements OnChanges, AfterViewInit, OnD
     public nocStore: Store<fromNocStore.State>,
     private readonly userService: UserService,
     private readonly featureToggleService: FeatureToggleService,
-    private readonly searchService: SearchService,
-    private readonly router: Router,
-    private readonly route: ActivatedRoute,
-    private renderer: Renderer2
+    private readonly searchService: SearchService
   ) { }
 
   public ngOnChanges(changes: SimpleChanges): void {
@@ -72,10 +65,6 @@ export class HmctsGlobalHeaderComponent implements OnChanges, AfterViewInit, OnD
     }
   }
 
-  public ngAfterViewInit(): void {
-    this.renderer.setProperty(this.caseReferenceEl.nativeElement,'innerHTML',"Hello Angular")
-  }
-
   public onEmitEvent(index: number): void {
     this.navigate.emit(this.navigation.items[index].emit);
   }
@@ -83,38 +72,6 @@ export class HmctsGlobalHeaderComponent implements OnChanges, AfterViewInit, OnD
   public onEmitSubMenu(menuItem: any) {
     if (menuItem.href === '/noc') {
       this.nocStore.dispatch(new fromNocStore.Reset());
-    }
-  }
-
-  public onSearchCase(caseReference): void {
-    // Populate a SearchParameters instance and persist via the SearchService
-    const searchParameters: SearchParameters = {
-      caseReferences: [caseReference],
-      CCDJurisdictionIds: null,
-      otherReferences: null,
-      fullName: null,
-      address: null,
-      postcode: null,
-      emailAddress: null,
-      dateOfBirth: null,
-      dateOfDeath: null
-    };
-
-    // Store the search parameters to session
-    this.searchService.storeState(SearchStatePersistenceKey.SEARCH_PARAMS, searchParameters);
-
-    this.searchSubscription$ = this.searchService.getResults().subscribe(result => {
-      if (result.resultInfo.casesReturned > 0) {
-        this.router.navigate([`/cases/case-details/${result.results[0].caseReference}`], { relativeTo: this.route });
-      } else {
-        this.router.navigate(['/search/noresults', NoResultsMessageId.NO_RESULTS_FROM_HEADER_SEARCH], { relativeTo: this.route });
-      }
-    });
-  }
-
-  public ngOnDestroy(): void {
-    if (this.searchSubscription$) {
-      this.searchSubscription$.unsubscribe();
     }
   }
