+++ src/app/shared/shared.module.ts
@@ -1,14 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
+import { WindowService } from '@hmcts/ccd-case-ui-toolkit';
+import { MediaViewerModule } from '@hmcts/media-viewer';
+import { ExuiCommonLibModule, GOV_UI_COMPONENTS } from '@hmcts/rpx-xui-common-lib';
 import * as fromAppComponents from '../components';
 import * as fromAppContainers from '../containers';
 import * as fromAppDirectives from '../directives';
 import { HealthCheckGuard } from './guards/health-check.guard';
 import { HealthCheckService } from './services/health-check.service';
-import { MediaViewerModule } from '@hmcts/media-viewer';
-import { WindowService } from '@hmcts/ccd-case-ui-toolkit';
-import { ExuiCommonLibModule, GOV_UI_COMPONENTS } from '@hmcts/rpx-xui-common-lib';
 
 /**
  * Shared Module
@@ -17,7 +18,13 @@ import { ExuiCommonLibModule, GOV_UI_COMPONENTS } from '@hmcts/rpx-xui-common-li
  */
 
 @NgModule( {
-  imports: [RouterModule, CommonModule, MediaViewerModule, ExuiCommonLibModule],
+  imports: [
+    RouterModule,
+    CommonModule,
+    MediaViewerModule,
+    ExuiCommonLibModule,
+    ReactiveFormsModule
+  ],
   declarations: [
     ...fromAppComponents.components,
     ...fromAppContainers.containers,
